Alterado poro de InsuranceApi para ExternalInsuranceCompany e movidos metodos que conectam com a API para model de InsuranceCompany
