+++ app/models/user.rb
@@ -11,18 +11,18 @@ class User < ApplicationRecord
   private
 
   def consult_insurance_company_api_for_email_validation
-    search_result = InsuranceApi.check_if_user_email_match_any_company(email)
+    search_result = InsuranceCompany.check_if_user_email_match_any_external_company(email)
     raise_company_error if search_result == []
     check_if_company_already_exists_locally(search_result)
   end
 
   def check_if_company_already_exists_locally(company)
-    found = InsuranceCompany.find_by external_insurance_company: company.id
-    if found.nil?
+    match = InsuranceCompany.find_by external_insurance_company: company.id
+    if match.nil?
       new_company = InsuranceCompany.create!(external_insurance_company: company.id)
       self.insurance_company_id = new_company.id
     else
-      self.insurance_company_id = found.id
+      self.insurance_company_id = match.id
     end
   end
