+++ spec/models/user_spec.rb
@@ -28,14 +28,14 @@ RSpec.describe User, type: :model do
 
   describe '#check_if_company_already_exists_locally' do
     it 'verifica se a seguradora já possui registro na aplicação, e adiciona id da seguradora ao usuário' do
-      company = InsuranceApi.new(
+      company = ExternalInsuranceCompany.new(
         id: 1,
         email_domain: 'petra@paolaseguros.com.br',
         company_status: 0,
         company_token: 'TOKENEXPIRADODESDE1999',
         token_status: 0
       )
-      allow(InsuranceApi).to receive(:check_if_user_email_match_any_company).and_return(company)
+      allow(InsuranceCompany).to receive(:check_if_user_email_match_any_external_company).and_return(company)
       FactoryBot.create(:insurance_company, external_insurance_company: company.id)
       user = User.create(email: 'petra@paolaseguros.com.br',
                          password: 'password',
@@ -48,14 +48,14 @@ RSpec.describe User, type: :model do
     end
 
     it 'seguradora não possuia resgistro na aplicação, salva a seguradora e adiciona o id da seguradora ao usuário' do
-      company = InsuranceApi.new(
+      company = ExternalInsuranceCompany.new(
         id: 1,
         email_domain: 'petra@paolaseguros.com.br',
         company_status: 0,
         company_token: 'TOKENEXPIRADODESDE1999',
         token_status: 0
       )
-      allow(InsuranceApi).to receive(:check_if_user_email_match_any_company).and_return(company)
+      allow(InsuranceCompany).to receive(:check_if_user_email_match_any_external_company).and_return(company)
       user = User.create(email: 'petra@paolaseguros.com.br',
                          password: 'password',
                          name: 'Petra',
